+++ app/models/channel.rb
@@ -1,6 +1,7 @@
 class Channel < ApplicationRecord
   include Rails.application.routes.url_helpers
   include Stripable
+  include ValidationErrorsNotifiable
 
   has_many :items, dependent: :destroy
   has_many :ownerships, dependent: :destroy
@@ -16,6 +17,7 @@ class Channel < ApplicationRecord
   validates :image_url, length: { maximum: 2083 }
 
   strip_before_save :title, :description
+  after_validation :notify_validation_errors
   after_create_commit { ChannelItemsUpdaterJob.perform_later(channel_id: self.id, mode: :all) }
 
   scope :not_stopped, -> { where.missing(:stopper) }
