+++ app/models/concerns/validation_errors_notifiable.rb
@@ -1,0 +1,11 @@
+module ValidationErrorsNotifiable
+  def notify_validation_errors
+    return if errors.empty?
+
+    content = [
+      "#{self.class} save failed: #{errors.full_messages.join(", ")}",
+      "```#{attributes.to_yaml}```",
+    ].join("\n")
+    DiscoPosterJob.perform_later(content: content)
+  end
+end
